+++ agent/src/test/groovy/sysmon/agent/AppTest.groovy
@@ -1,0 +1,7 @@
+package sysmon.agent
+
+import spock.lang.Specification
+
+class AppTest extends Specification {
+
+}